pull accessibility mode out of editor status (#168103)
